refactor: break large files into smaller * code otherwise the same though there seems to a lot of repitition
