+++ src/config/probuff.config.ts
@@ -1,0 +1,9 @@
+import _m0 from "protobufjs/minimal";
+import Long from "long";
+
+if (_m0.util.Long !== Long) {
+  _m0.util.Long = Long as any;
+  _m0.configure();
+}
+
+export default _m0
