+++ src/kima/tx/cancel.ts
@@ -1,0 +1,147 @@
+import _m0 from "../../config/probuff.config";
+import { DeepPartial, Exact, isSet } from "../../utils";
+
+export interface MsgCancelTransaction {
+  creator: string;
+  transactionId: string;
+}
+
+export interface MsgCancelTransactionResponse {
+  code: string;
+  msg: string;
+}
+
+function createBaseMsgCancelTransaction(): MsgCancelTransaction {
+  return { creator: "", transactionId: "" };
+}
+
+export const MsgCancelTransaction = {
+  encode(
+    message: MsgCancelTransaction,
+    writer: _m0.Writer = _m0.Writer.create()
+  ): _m0.Writer {
+    if (message.creator !== "") {
+      writer.uint32(10).string(message.creator);
+    }
+    if (message.transactionId !== "") {
+      writer.uint32(18).string(message.transactionId);
+    }
+    return writer;
+  },
+
+  decode(
+    input: _m0.Reader | Uint8Array,
+    length?: number
+  ): MsgCancelTransaction {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseMsgCancelTransaction();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.creator = reader.string();
+          break;
+        case 2:
+          message.transactionId = reader.string();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): MsgCancelTransaction {
+    return {
+      creator: isSet(object.creator) ? String(object.creator) : "",
+      transactionId: isSet(object.transactionId)
+        ? String(object.transactionId)
+        : "",
+    };
+  },
+
+  toJSON(message: MsgCancelTransaction): unknown {
+    const obj: any = {};
+    message.creator !== undefined && (obj.creator = message.creator);
+    message.transactionId !== undefined &&
+      (obj.transactionId = message.transactionId);
+    return obj;
+  },
+
+  fromPartial<I extends Exact<DeepPartial<MsgCancelTransaction>, I>>(
+    object: I
+  ): MsgCancelTransaction {
+    const message = createBaseMsgCancelTransaction();
+    message.creator = object.creator ?? "";
+    message.transactionId = object.transactionId ?? "";
+    return message;
+  },
+};
+
+function createBaseMsgCancelTransactionResponse(): MsgCancelTransactionResponse {
+  return { code: "", msg: "" };
+}
+
+export const MsgCancelTransactionResponse = {
+  encode(
+    message: MsgCancelTransactionResponse,
+    writer: _m0.Writer = _m0.Writer.create()
+  ): _m0.Writer {
+    if (message.code !== "") {
+      writer.uint32(10).string(message.code);
+    }
+    if (message.msg !== "") {
+      writer.uint32(18).string(message.msg);
+    }
+    return writer;
+  },
+
+  decode(
+    input: _m0.Reader | Uint8Array,
+    length?: number
+  ): MsgCancelTransactionResponse {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseMsgCancelTransactionResponse();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.code = reader.string();
+          break;
+        case 2:
+          message.msg = reader.string();
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): MsgCancelTransactionResponse {
+    return {
+      code: isSet(object.code) ? String(object.code) : "",
+      msg: isSet(object.msg) ? String(object.msg) : "",
+    };
+  },
+
+  toJSON(message: MsgCancelTransactionResponse): unknown {
+    const obj: any = {};
+    message.code !== undefined && (obj.code = message.code);
+    message.msg !== undefined && (obj.msg = message.msg);
+    return obj;
+  },
+
+  fromPartial<I extends Exact<DeepPartial<MsgCancelTransactionResponse>, I>>(
+    object: I
+  ): MsgCancelTransactionResponse {
+    const message = createBaseMsgCancelTransactionResponse();
+    message.code = object.code ?? "";
+    message.msg = object.msg ?? "";
+    return message;
+  },
+};
