+++ src/kima/tx/index.ts
@@ -1,0 +1,1 @@
+export const protobufPackage = "kimablockchain.transaction";
