+++ src/kima/tx/response.ts
@@ -1,0 +1,83 @@
+import _m0 from "../../config/probuff.config";
+import { DeepPartial, Exact, isSet, longToNumber } from "../../utils";
+
+export interface MsgRequestTransactionResponse {
+  code: string;
+  msg: string;
+  txId: number;
+}
+
+function createBaseMsgRequestTransactionResponse(): MsgRequestTransactionResponse {
+  return { code: "", msg: "", txId: 0 };
+}
+
+export const MsgRequestTransactionResponse = {
+  encode(
+    message: MsgRequestTransactionResponse,
+    writer: _m0.Writer = _m0.Writer.create()
+  ): _m0.Writer {
+    if (message.code !== "") {
+      writer.uint32(10).string(message.code);
+    }
+    if (message.msg !== "") {
+      writer.uint32(18).string(message.msg);
+    }
+    if (message.txId !== 0) {
+      writer.uint32(24).uint64(message.txId);
+    }
+    return writer;
+  },
+
+  decode(
+    input: _m0.Reader | Uint8Array,
+    length?: number
+  ): MsgRequestTransactionResponse {
+    const reader = input instanceof _m0.Reader ? input : new _m0.Reader(input);
+    let end = length === undefined ? reader.len : reader.pos + length;
+    const message = createBaseMsgRequestTransactionResponse();
+    while (reader.pos < end) {
+      const tag = reader.uint32();
+      switch (tag >>> 3) {
+        case 1:
+          message.code = reader.string();
+          break;
+        case 2:
+          message.msg = reader.string();
+          break;
+        case 3:
+          message.txId = longToNumber(reader.uint64() as Long);
+          break;
+        default:
+          reader.skipType(tag & 7);
+          break;
+      }
+    }
+    return message;
+  },
+
+  fromJSON(object: any): MsgRequestTransactionResponse {
+    return {
+      code: isSet(object.code) ? String(object.code) : "",
+      msg: isSet(object.msg) ? String(object.msg) : "",
+      txId: isSet(object.txId) ? Number(object.txId) : 0,
+    };
+  },
+
+  toJSON(message: MsgRequestTransactionResponse): unknown {
+    const obj: any = {};
+    message.code !== undefined && (obj.code = message.code);
+    message.msg !== undefined && (obj.msg = message.msg);
+    message.txId !== undefined && (obj.txId = Math.round(message.txId));
+    return obj;
+  },
+
+  fromPartial<I extends Exact<DeepPartial<MsgRequestTransactionResponse>, I>>(
+    object: I
+  ): MsgRequestTransactionResponse {
+    const message = createBaseMsgRequestTransactionResponse();
+    message.code = object.code ?? "";
+    message.msg = object.msg ?? "";
+    message.txId = object.txId ?? 0;
+    return message;
+  },
+};
